+++ taxel-gui/src/ui/header.rs
@@ -5,8 +5,7 @@ use crate::{
     TaxelApp,
 };
 use eframe::egui::{
-    text::LayoutJob, vec2, Align, Button, Color32, Id, Layout, Modal, Shape, TextEdit, TextFormat,
-    Ui, Widget,
+    text::LayoutJob, vec2, Align, Button, Color32, Layout, Shape, TextEdit, TextFormat, Ui,
 };
 use rfd::FileDialog;
 
@@ -25,7 +24,7 @@ pub fn draw_header(ui: &mut Ui, app: &mut TaxelApp) {
                     Ok(copied_path) => {
                         app.register_report(&copied_path);
                         app.refresh_reports();
-                        app::load_report(app, copied_path, ui.ctx().clone());
+                        app::load_report(app, copied_path, ui.ctx().clone(), false);
                     }
                     Err(err) => {
                         app.diagnostics.push(AppDiagnostic::new_error(
@@ -84,104 +83,6 @@ pub fn draw_header(ui: &mut Ui, app: &mut TaxelApp) {
             draw_zoom_toolbar(ui, &mut app.settings.zoom_input);
         });
     });
-}
-
-/// Draws the delete confirmation modal when the user clicks "Delete report".
-pub fn draw_delete_modal(ui: &mut Ui, app: &mut TaxelApp) {
-    let filename = app
-        .report
-        .as_ref()
-        .and_then(|report| report.path.file_name())
-        .and_then(|file_name| file_name.to_str())
-        .unwrap_or("this report");
-
-    let mut cancel = false;
-    let mut confirm = false;
-
-    let modal = Modal::new(Id::new("delete_report_modal")).show(ui.ctx(), |ui| {
-        ui.heading("Move report to Trash?");
-        ui.add_space(4.0);
-        ui.label(filename);
-        ui.add_space(8.0);
-        ui.horizontal(|ui| {
-            if ui.button("Cancel").clicked() {
-                cancel = true;
-            }
-            if ui.button("Move to Trash").clicked() {
-                confirm = true;
-            }
-        });
-    });
-
-    if modal.should_close() || cancel {
-        app.show_delete_modal = false;
-    }
-
-    if confirm {
-        app.show_delete_modal = false;
-        app::delete_report(app);
-    }
-}
-
-/// Draws the send report modal when the user clicks "Send report".
-pub fn draw_send_modal(ui: &mut Ui, app: &mut TaxelApp) {
-    let mut cancel = false;
-    let mut confirm = false;
-
-    let modal = Modal::new(Id::new("send_report_modal")).show(ui.ctx(), |ui| {
-        ui.heading("Send report");
-        ui.add_space(8.0);
-
-        ui.horizontal(|ui| {
-            ui.label("Certificate");
-            let path_label = app
-                .send_certificate_path
-                .as_deref()
-                .and_then(|p| p.to_str())
-                .unwrap_or("No file selected");
-            ui.label(path_label);
-            if ui.button("Browse…").clicked() {
-                if let Some(path) = FileDialog::new()
-                    .add_filter("PFX certificate", &["pfx", "p12"])
-                    .add_filter("All", &["*"])
-                    .pick_file()
-                {
-                    app.send_certificate_path = Some(path);
-                }
-            }
-        });
-
-        ui.add_space(4.0);
-
-        ui.horizontal(|ui| {
-            ui.label("Password");
-            TextEdit::singleline(&mut app.send_password)
-                .password(true)
-                .desired_width(180.0)
-                .ui(ui);
-        });
-
-        ui.add_space(8.0);
-
-        ui.horizontal(|ui| {
-            if ui.button("Cancel").clicked() {
-                cancel = true;
-            }
-            let can_send = app.send_certificate_path.is_some() && !app.send_password.is_empty();
-            if ui.add_enabled(can_send, Button::new("Send")).clicked() {
-                confirm = true;
-            }
-        });
-    });
-
-    if modal.should_close() || cancel {
-        app.show_send_modal = false;
-    }
-
-    if confirm {
-        app.show_send_modal = false;
-        app::send_report(app);
-    }
 }
 
 /// Draws a summary of errors and warnings in the header. Clicking on the
