+++ taxel-gui/src/ui/mod.rs
@@ -1,11 +1,15 @@
 mod diagnostic_panel;
 mod header;
+mod modals;
 mod report_list_view;
 mod report_view;
 mod widgets;
 
 pub use diagnostic_panel::draw_error_panel;
-pub use header::{draw_delete_modal, draw_header, draw_send_modal};
+pub use header::draw_header;
+pub use modals::{
+    draw_delete_modal, draw_download_modal, draw_send_modal, draw_unsaved_changes_modal,
+};
 pub use report_list_view::draw_report_list;
 pub use report_view::{
     search_overlay::{draw_search_results_overlay, highlight_row},
@@ -13,4 +17,3 @@ pub use report_view::{
     table::draw_table,
     toolbar::{draw_toolbar, EditAction},
 };
-pub use widgets::draw_unsaved_changes_modal;
